+++ packages/material-ui/src/styles/createTransitions.d.ts
