+++ packages/material-ui/src/styles/index.js
@@ -7,7 +7,7 @@ export { getUnit as unstable_getUnit, toUnitless as unstable_toUnitless } from '
 export { default as makeStyles } from './makeStyles';
 export { default as responsiveFontSizes } from './responsiveFontSizes';
 export { default as styled } from './styled';
-export { duration, easing } from './transitions';
+export { duration, easing } from './createTransitions';
 export { default as useTheme } from './useTheme';
 export { default as unstable_useThemeProps } from './useThemeProps';
 export { default as withStyles } from './withStyles';
